Remove usdlib package, usdQt variantSets module
The variantSet model isn't really at a point where it feels suitable
for general consumption, and the bulk of usdlib is currently more of a
support package for it than a set of general-purpose utilities.
